+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # 1.0.4
 
 - `Entity.getKeyValueAsBool()` parses non-zero integer values as true and zero integer values as false.
-- Added tests to catch further parsing errors.
-
+- Added new tests to catch further parsing errors. Further updates will improve testing suite.
+- `getKeyValueAsBool()` now correctly sets `caseSenitive` param in `bool.parse()` to `true`.
+- Added support for getKeyValueAsBool() to return true for non-zero `int` values.
+  
 # 1.0.3
 
 - Missed or.unquote() for Element.getKeyValue() which affected all the other keyval getters. Fixed.